pyproject.toml
[tool.poetry]
name = "fastapi-ex-user"
version = "0.1.0"
description = ""
authors = ["Abhishek koserwal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-keycloak = "^4.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
